src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/offset-parent-does-not-leak-ua-shadow-expected.txt

CONSOLE ERROR: line 33: Uncaught TypeError: host.createShadowRoot is not a function
offsetParent should not leak nodes in user agent Shadow DOM.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
